% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps.R
\name{pps}
\alias{pps}
\title{Get list of paths between two nodes and their PPS}
\usage{
pps(P, i, j, K = 5, prec = TRUE, use.names = TRUE)
}
\arguments{
\item{P}{A precision matrix or partial correlation matrix.}

\item{i, j}{Node numbers. Should be positive integers.}

\item{K}{The maximum path length to search up to. Defaults to 5.}

\item{prec}{If \code{TRUE}, indicates that \code{P} is a precision matrix.
If \code{FALSE}, \code{P} is taken to be a partial correlation matrix.
Defaults to TRUE.}

\item{use.names}{Whether or not to ouput paths using node names. Defaults to \code{TRUE}.
If the precision/partial correlation matrix does not have column names,
indices will be used.}
}
\value{
A list with the following elements:
\item{path}{A list of paths between nodes \code{i} and \code{j} in order of descending PPS.}
\item{pps}{A list of the PPS values for the paths in \code{path}.}
\item{gamma}{The individual contributions of each path in \code{path} on the correlation scale.}
}
\description{
For a given precision matrix or partial correlation matrix and a given
pair of nodes, computes all paths between those nodes up to a certain length
and their PPS
}
