% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.PPS.App.R
\name{run.PPS.App}
\alias{run.PPS.App}
\title{Launch RShiny app for PPS analysis}
\usage{
run.PPS.App()
}
\description{
Launches an RShiny app to for PPS analysis. The
app is essentially a wrapper for the \code{pps}
function, with some additional visualizations.
}
