\name{sizesok}
\alias{sizesok}
\title{Check that unit sizes are not too big}
\description{See user's guide}
\usage{sizesok(size,n)}
\arguments{
	\item{size}{A vector of the sizes of the units in the population}
	\item{n}{The sample size}
}
\value{Returns the number of 'bad' units}
\note{See the user's guide, pps-ug.pdf, for more information.}
\keyword{survey}

