\name{stratumsizes}
\alias{stratumsizes}
\title{Compute size of each stratum}
\description{Given a vector of sorted stratum indicators, returns the number of units in each stratum}
\usage{stratumsizes(stratum)}
\arguments{
	\item{stratum}{A vector of sorted stratum indicators}
}
\value{Returns the number of units in each stratum}
\note{See the user's guide, pps.pdf, for more information.}
\examples{
}
\keyword{survey}

